# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SecureProfit Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feedback360/{id}:
    patch:
      operationId: submitFeedback360
      tags: [feedback360]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitFeedback360Body" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feedback360" }
  /projects/{id}/feedback360:
    get:
      operationId: listProjectFeedback360
      tags: [feedback360]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Feedback360" }
...
components:
  schemas:
...
    Feedback360:
      type: object
      required: [id, projectId, projectCode, projectName, reviewerId, reviewerName, subjectId, subjectName, status, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        projectCode: { type: string }
        projectName: { type: string }
        reviewerId: { type: string }
        reviewerName: { type: string }
        reviewerRole: { type: string }
        subjectId: { type: string }
        subjectName: { type: string }
        subjectRole: { type: string }
        rating: { type: integer, nullable: true }
        comment: { type: string, nullable: true }
        status: { type: string, enum: [PENDING, SUBMITTED] }
        submittedAt: { type: string, nullable: true }
        createdAt: { type: string }
    SubmitFeedback360Body:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
...