# lib/api-zod/src/generated/types/feedback360.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SecureProfit Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Feedback360Status } from "./feedback360Status";

export interface Feedback360 {
  id: string;
  projectId: string;
  projectCode: string;
  projectName: string;
  reviewerId: string;
  reviewerName: string;
  reviewerRole?: string;
  subjectId: string;
  subjectName: string;
  subjectRole?: string;
  rating?: number | null;
  comment?: string | null;
  status: Feedback360Status;
  submittedAt?: string | null;
  createdAt: string;
}